requirements.txt
SQLAlchemy~=2.0.31
bcrypt~=4.1.3
Flask~=3.0.3
mysql~=0.0.3
mysql-connector-python~=9.0.0
PyJWT~=2.8.0
pandas~=2.2.2
matplotlib~=3.9.1
crypto~=1.4.1
setuptools~=71.1.0